aklite: move the client code from helper files
